Drop the py27/swig event loop
